#176-binding image 부분을 util로 이동
